+++ src/main/java/pl/edu/agh/awi/Application.java
@@ -22,7 +22,6 @@ public class Application implements CommandLineRunner {
     }
 
     @Override
-    @Transactional
     public void run(String[] args) throws Exception {
         loaders.forEach(e -> e.loadData());
     }
