--- src/main/java/pl/edu/agh/awi/downloader/exceptions/FlightTaskException.java
@@ -1,9 +1,0 @@
-package pl.edu.agh.awi.downloader.exceptions;
-
-public class FlightTaskException extends RuntimeException {
-
-    public FlightTaskException(String s) {
-        super(s);
-    }
-
-}
