+++ src/main/java/pl/edu/agh/awi/downloader/flights/flightDetails/data/FlightDetailsResponse.java
@@ -49,6 +49,18 @@ public class FlightDetailsResponse {
 
     private final String originTimezoneName;
 
+    private final String realIata;
+
+    private final String realCity;
+
+    private final List<Double> realCoords;
+
+    private final String realTimezoneCode;
+
+    private final double realTimezoneOffset;
+
+    private final String realTimezoneName;
+
     private final String airline;
 
     private final String aircraft;
@@ -67,28 +79,39 @@ public class FlightDetailsResponse {
 
     private final String imageLink;
 
+    private final String imageLarge;
+
+    private final String imagelinkLarge;
+
     private final String copyright;
+
+    private final String copyrightLarge;
 
     private final Integer firstTimestamp;
 
     public FlightDetailsResponse(@JsonProperty("q") String fightId, @JsonProperty("flight") String flightDesignator,
                                  @JsonProperty("status") String status,
                                  @JsonProperty("dep_schd") @JsonDeserialize(using = CustomDateDeserializer.class) Date std,
-                                 @JsonProperty("arr_schd") @JsonDeserialize(using = CustomDateDeserializer.class)Date sta,
-                                 @JsonProperty("departure") @JsonDeserialize(using = CustomDateDeserializer.class)Date atd,
-                                 @JsonProperty("arrival") @JsonDeserialize(using = CustomDateDeserializer.class)Date ata,
-                                 @JsonProperty("eta") @JsonDeserialize(using = CustomDateDeserializer.class)Date eta,
+                                 @JsonProperty("arr_schd") @JsonDeserialize(using = CustomDateDeserializer.class) Date sta,
+                                 @JsonProperty("departure") @JsonDeserialize(using = CustomDateDeserializer.class) Date atd,
+                                 @JsonProperty("arrival") @JsonDeserialize(using = CustomDateDeserializer.class) Date ata,
+                                 @JsonProperty("eta") @JsonDeserialize(using = CustomDateDeserializer.class) Date eta,
                                  @JsonProperty("to_iata") String destinationIata, @JsonProperty("to_city") String destinationCity,
                                  @JsonProperty("to_pos") List<Double> destinationCoords, @JsonProperty("to_tz_code") String destinationTimezoneCode,
                                  @JsonProperty("to_tz_offset") double destinationTimezoneOffset, @JsonProperty("to_tz_name") String destinationTimezoneName,
                                  @JsonProperty("from_iata") String originIata, @JsonProperty("from_city") String originCity,
                                  @JsonProperty("from_pos") List<Double> originCoords, @JsonProperty("from_tz_code") String originTimezoneCode,
                                  @JsonProperty("from_tz_offset") double originTimezoneOffset, @JsonProperty("from_tz_name") String originTimezoneName,
+                                 @JsonProperty("real_to_iata") String realIata, @JsonProperty("real_to_city") String realCity,
+                                 @JsonProperty("real_to_pos") List<Double> realCoords, @JsonProperty("real_to_tz_code") String realTimezoneCode,
+                                 @JsonProperty("real_to_tz_offset") double realTimezoneOffset, @JsonProperty("real_to_tz_name") String realTimezoneName,
                                  @JsonProperty("airline") String airline, @JsonProperty("aircraft") String aircraft,
                                  @JsonProperty("trail") List<Double> trail, @JsonProperty("snapshot_id") String snapshotId,
                                  @JsonProperty("airline_url") String airlineUrl, @JsonProperty("sideview") String sideview,
                                  @JsonProperty("image") String image, @JsonProperty("imagelink") String imageLink,
-                                 @JsonProperty("copyright") String copyright, @JsonProperty("first_timestamp") Integer firstTimestamp) {
+                                 @JsonProperty("image_large") String imageLarge, @JsonProperty("imagelink_large") String imagelinkLarge,
+                                 @JsonProperty("copyright") String copyright, @JsonProperty("copyright_large") String copyrightLarge,
+                                 @JsonProperty("first_timestamp") Integer firstTimestamp) {
         this.fightId = fightId;
         this.flightDesignator = flightDesignator;
         this.status = status;
@@ -109,6 +132,12 @@ public class FlightDetailsResponse {
         this.originTimezoneCode = originTimezoneCode;
         this.originTimezoneOffset = originTimezoneOffset;
         this.originTimezoneName = originTimezoneName;
+        this.realIata = realIata;
+        this.realCity = realCity;
+        this.realCoords = realCoords;
+        this.realTimezoneCode = realTimezoneCode;
+        this.realTimezoneOffset = realTimezoneOffset;
+        this.realTimezoneName = realTimezoneName;
         this.airline = airline;
         this.aircraft = aircraft;
         this.trail = trail;
@@ -119,7 +148,10 @@ public class FlightDetailsResponse {
         this.sideview = sideview;
         this.image = image;
         this.imageLink = imageLink;
+        this.imageLarge = imageLarge;
+        this.imagelinkLarge = imagelinkLarge;
         this.copyright = copyright;
+        this.copyrightLarge = copyrightLarge;
         this.firstTimestamp = firstTimestamp;
     }
