+++ src/main/java/pl/edu/agh/awi/scheduler/exception/SchedulerException.java
@@ -1,0 +1,9 @@
+package pl.edu.agh.awi.scheduler.exception;
+
+public class SchedulerException extends RuntimeException {
+
+    public SchedulerException(String s) {
+        super(s);
+    }
+
+}
